comment_crawler/requirements.txt
scrapy_splash==0.7.2